[4142] Change GIAS namespace
