+++ app/services/gias/importer/mentor_at_school_periods/overlapping.rb
@@ -1,5 +1,5 @@
 module GIAS
-  module Schools
+  class Importer
     module MentorAtSchoolPeriods
       class Overlapping
         attr_reader :teacher, :schools, :groups
