+++ spec/services/gias/importer/mentor_at_school_periods/overlapping_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe GIAS::Schools::MentorAtSchoolPeriods::Overlapping do
+RSpec.describe GIAS::Importer::MentorAtSchoolPeriods::Overlapping do
   subject(:groups) { described_class.find(teacher:, schools:) }
 
   let(:teacher) { FactoryBot.create(:teacher) }
