+++ spec/services/gias/importer/replace_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe GIAS::Schools::Replace do
+RSpec.describe GIAS::Importer::Replace do
   describe "#replace!" do
     subject(:service) { described_class.new(gias_school).replace! }
