+++ spec/services/gias/importer/school_partnerships/transfer_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe GIAS::Schools::SchoolPartnerships::Transfer do
+RSpec.describe GIAS::Importer::SchoolPartnerships::Transfer do
   subject(:service) do
     described_class.call(
       predecessor_school_partnership:,
